hardcode round constants
Make the round function take the number of rounds as an integer instead
of an arbitrary slice of round constants. The old way was cute but it
provided more flexability than is needed.

The round constants are very simple (they can be computed as
r_i = 0xf0 - 15*(i + (12-b)), where i is the round number and b is the
number of rounds). I tried a couple ways of computing them on the fly
but it turned out to be slower than using a table. I guess the increased
register pressure outweighed the memory latency cost, or something.